feat: expose aggregate metrics, enable saving
