+++ evals/eval_framework/eval_config.py
@@ -22,6 +22,9 @@ class EvalConfig(BaseSettings):
     evaluation_metrics: List[str] = ["correctness", "EM", "f1"]
     deepeval_model: str = "gpt-4o-mini"
 
+    # Metrics params
+    calculate_metrics: bool = True
+
     # Visualization
     dashboard: bool = True
 
@@ -29,6 +32,7 @@ class EvalConfig(BaseSettings):
     questions_path: str = "questions_output.json"
     answers_path: str = "answers_output.json"
     metrics_path: str = "metrics_output.json"
+    aggregate_metrics_path: str = "aggregate_metrics.json"
     dashboard_path: str = "dashboard.html"
 
     model_config = SettingsConfigDict(env_file=".env", extra="allow")
@@ -43,10 +47,12 @@ class EvalConfig(BaseSettings):
             "evaluating_answers": self.evaluating_answers,
             "evaluation_engine": self.evaluation_engine,
             "evaluation_metrics": self.evaluation_metrics,
+            "calculate_metrics": self.calculate_metrics,
             "dashboard": self.dashboard,
             "questions_path": self.questions_path,
             "answers_path": self.answers_path,
             "metrics_path": self.metrics_path,
+            "aggregate_metrics_path": self.aggregate_metrics_path,
             "dashboard_path": self.dashboard_path,
             "deepeval_model": self.deepeval_model,
             "task_getter_type": self.task_getter_type,
